+++ VWEB/VWEB.csproj
@@ -152,6 +152,7 @@
     <Compile Include="App_Start\RouteConfig.cs" />
     <Compile Include="Controllers\AlunosController.cs" />
     <Compile Include="Controllers\HomeController.cs" />
+    <Compile Include="Controllers\LoginController.cs" />
     <Compile Include="Controllers\MensagensController.cs" />
     <Compile Include="Controllers\PostagensController.cs" />
     <Compile Include="Controllers\ResponsaveisController.cs" />
@@ -258,8 +259,6 @@
     <Content Include="Views\_ViewStart.cshtml" />
     <Content Include="Views\Shared\Error.cshtml" />
     <Content Include="Views\Shared\_Layout.cshtml" />
-    <Content Include="Views\Home\About.cshtml" />
-    <Content Include="Views\Home\Contact.cshtml" />
     <Content Include="Views\Home\Index.cshtml" />
     <Content Include="Views\Alunos\Create.cshtml" />
     <Content Include="Views\Alunos\Delete.cshtml" />
@@ -302,6 +301,7 @@
     <Content Include="Views\Postagens\Details.cshtml" />
     <Content Include="Views\Postagens\Edit.cshtml" />
     <Content Include="Views\Postagens\Index.cshtml" />
+    <Content Include="Views\Login\Index.cshtml" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Uploads\" />
