+++ VWEB/Views/Mensagens/Create.cshtml
@@ -36,7 +36,7 @@
             </div>
 
             <div class="md-form md-outline">
-                @Html.TextAreaFor(model => model.Texto, new { @class = "md-textarea form-control validate", rows = "3", minlength = "30", maxlength = "600" })
+                @Html.TextAreaFor(model => model.Texto, new { @class = "md-textarea form-control validate", rows = "3", minlength = "40", maxlength = "600" })
                 @Html.LabelFor(model => model.Texto, htmlAttributes: new { data_error = "Inválido", data_success = "" })
             </div>
