+++ VWEB/Views/Shared/_TemplatePrincipal.cshtml
@@ -164,10 +164,6 @@
                     <span class="badge badge-pill badge-success notification">7</span>
                 </a>
                 <a href="#">
-                    <i class="fa fa-cog"></i>
-                    <span class="badge-sonar"></span>
-                </a>
-                <a href="#">
                     <i class="fa fa-power-off"></i>
                 </a>
             </div>
